moving nav tests around to be more efficient
