+++ e2e/pages/shared-lending-platform/unauthenticated-homepage/NavMenu.spec.ts
@@ -1,0 +1,8 @@
+import { expect, test } from '@playwright/test';
+
+test('Unauthenticated homepage: Validate Nav Menu', async ({ page }) => {
+  await test.step('Verify nav options', async () => {
+    const navContainer = page.locator('nav#nav-links');
+    await expect(navContainer).not.toBeVisible();
+  });
+});
